requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.1.3
executing==2.0.1
extra-streamlit-components==0.1.60
fonttools==4.45.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.18.0
jedi==0.19.1
Jinja2==3.1.2
jovian==0.2.47
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
openai==0.28.0
packaging==23.2
pandas==2.1.2
parquet==1.3.1
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==4.0.0
plotly==5.18.0
ply==3.11
prompt-toolkit==3.0.41
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pypdfium2==4.24.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.28.0
streamlit-authenticator==0.1.5
svgwrite==1.4.3
tenacity==8.2.3
thriftpy2==0.4.17
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
Tree==0.2.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uuid==1.30
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
weaviate-client==3.25.3
yarl==1.9.2
zipp==3.17.0
